Feat: 유저 서비스에 세션을 통한 검증 추가

+++ src/main/java/com/example/newsfeed/user/application/service/UserService.java
@@ -3,6 +3,7 @@ package com.example.newsfeed.user.application.service;
 import com.example.newsfeed.global.common.exception.ErrorDetail;
 import com.example.newsfeed.user.application.converter.UserConverter;
 import com.example.newsfeed.user.dto.request.DeleteUserRequestDto;
+import com.example.newsfeed.user.dto.request.SearchUserRequestDto;
 import com.example.newsfeed.user.dto.request.UpdateUserPasswordRequestDto;
 import com.example.newsfeed.user.dto.request.UpdateUserRequestDto;
 import com.example.newsfeed.user.dto.response.GetAllUsersResponseDto;
@@ -10,9 +11,12 @@ import com.example.newsfeed.user.dto.response.GetUserResponseDto;
 import com.example.newsfeed.user.entity.User;
 import com.example.newsfeed.user.exception.InvalidPasswordException;
 import com.example.newsfeed.user.exception.PasswordSameAsOldException;
+import com.example.newsfeed.user.exception.UserAccessDeniedException;
 import com.example.newsfeed.user.exception.UserNotFoundException;
 import com.example.newsfeed.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -28,21 +32,28 @@ public class UserService {
     private final UserRepository userRepository;
     private final PasswordEncoder passwordEncoder;
 
+    // 유저 단건 조회
     @Transactional(readOnly = true)
     public GetUserResponseDto findUser(Long userId) {
         User findUser = findUserById(userId);
         return UserConverter.toResponse(findUser);
     }
 
+    // 키워드로 유저 전체 조회
     @Transactional(readOnly = true)
-    public List<GetAllUsersResponseDto> findUsers() {
-        List<User> users = userRepository.findAllUsers();
+    public Page<GetAllUsersResponseDto> findUsers(SearchUserRequestDto dto, Pageable pageable) {
+        Page<User> users = userRepository.findAllUsersByKeyword(dto.getKeyword(), pageable);
         return UserConverter.toResponse(users);
     }
 
+    // 유저 프로필 수정
     @Transactional
-    public Long updateUserProfile(Long userId, UpdateUserRequestDto dto) {
-        //TODO: 회원 인증 로직 구현 이후 로그인 유저와 저장된 유저 간의 검증 필요
+    public Long updateUserProfile(
+        Long userId,
+        UpdateUserRequestDto dto,
+        Long sessionUserId
+    ) {
+        checkUserPermission(userId, sessionUserId);
         User findUser = findUserById(userId);
 
         validatePassword(dto.getPassword(), findUser.getPassword());
@@ -60,9 +71,14 @@ public class UserService {
         return findUser.getId();
     }
 
+    // 유저 비밀번호 수정
     @Transactional
-    public Long updateUserPassword(Long userId, UpdateUserPasswordRequestDto dto) {
-        //TODO: 회원 인증 로직 구현 이후 로그인 유저와 저장된 유저 간의 검증 필요
+    public Long updateUserPassword(
+        Long userId,
+        UpdateUserPasswordRequestDto dto,
+        Long sessionUserId
+    ) {
+        checkUserPermission(userId, sessionUserId);
         User findUser = findUserById(userId);
 
         validatePassword(dto.getPassword(), findUser.getPassword());
@@ -77,20 +93,34 @@ public class UserService {
         return findUser.getId();
     }
 
+    // 유저 탈퇴
     @Transactional
-    public void deleteUser(Long userId, DeleteUserRequestDto dto) {
-        //TODO: 회원 인증 로직 구현 이후 로그인 유저와 저장된 유저 간의 검증 필요
+    public void deleteUser(
+        Long userId,
+        DeleteUserRequestDto dto,
+        Long sessionUserId
+    ) {
+        checkUserPermission(userId, sessionUserId);
         User findUser = findUserById(userId);
         validatePassword(dto.getPassword(), findUser.getPassword());
+        //TODO: 피드 및 댓글까지 전부 삭제해야 함
         userRepository.deleteUserById(userId);
     }
 
     public User findUserById(Long userId) {
-        //TODO: public으로 열지 않고 다른 도메인의 서비스에서도 호출하도록 로직을 바꿀 수 있는지 확인
         return userRepository.findUserById(userId)
             .orElseThrow(() -> new UserNotFoundException(List.of(
                 new ErrorDetail(USER_NOT_FOUND, null, USER_NOT_FOUND.getMessage())
             )));
+    }
+
+    // 로그인한 유저의 리소스 접근 허용 여부 확인
+    private void checkUserPermission(Long userId, Long sessionUserId) {
+        if (!userId.equals(sessionUserId)) {
+            throw new UserAccessDeniedException(List.of(
+                new ErrorDetail(USER_ACCESS_DENIED, null, USER_ACCESS_DENIED.getMessage())
+            ));
+        }
     }
 
     private void validatePassword(String inputPassword, String storedPassword) {
